feat: extract LoginForm component from login page
Closes #978

- Created LoginForm as a reusable presentational component
- Moved form logic, validation, and error handling to LoginForm
- Updated login page to use LoginForm component
- Improves testability and component reusability
