+++ packages/app/src/app/[locale]/auth/login/page.tsx
@@ -1,54 +1,9 @@
 "use client";
 
-import { useState } from "react";
 import Link from "next/link";
-import { useRouter, useSearchParams } from "next/navigation";
-import { useForm } from "react-hook-form";
-import { zodResolver } from "@hookform/resolvers/zod";
-import { Loader2 } from "lucide-react";
-import { loginSchema, type LoginInput, authApi } from "@/lib/auth";
-import { useAuth } from "@/context/AuthContext";
-import type { AuthUser } from "@/context/AuthContext";
-import FormField from "@/components/FormField";
-import { cn } from "@/lib/utils";
+import LoginForm from "@/components/LoginForm";
 
 export default function LoginPage() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const redirect = searchParams.get("redirect") ?? "/workers";
-  const { login } = useAuth();
-  const [apiError, setApiError] = useState<string | null>(null);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isSubmitting, touchedFields, dirtyFields },
-  } = useForm<LoginInput>({
-    resolver: zodResolver(loginSchema),
-    mode: "onChange",
-  });
-
-  const isValid = (field: keyof LoginInput) =>
-    dirtyFields[field] && !errors[field];
-
-  const onSubmit = async (data: LoginInput) => {
-    setApiError(null);
-    try {
-      const res = await authApi.login(data);
-      login(res.data as AuthUser, res.token);
-      router.push(redirect);
-    } catch (err: unknown) {
-      setApiError(err instanceof Error ? err.message : "Login failed");
-    }
-  };
-
-  const inputClass = (hasError?: boolean, valid?: boolean) =>
-    cn(
-      "w-full rounded-lg border px-3 py-2.5 pr-9 text-sm outline-none focus:ring-2 focus:ring-blue-500 transition-colors",
-      hasError && "border-red-400 focus:ring-red-300",
-      valid && !hasError && "border-green-400 focus:ring-green-300"
-    );
-
   return (
     <div className="flex min-h-screen items-center justify-center bg-gray-50 px-4">
       <div className="w-full max-w-md rounded-2xl border bg-white p-8 shadow-sm">
@@ -74,66 +29,7 @@ export default function LoginPage() {
           <div className="h-px flex-1 bg-gray-200" />
         </div>
 
-        <form onSubmit={handleSubmit(onSubmit)} className="flex flex-col gap-4" noValidate>
-          {apiError && (
-            <div className="rounded-lg bg-red-50 px-4 py-3 text-sm text-red-600">
-              {apiError}
-            </div>
-          )}
-
-          <FormField
-            label="Email"
-            id="email"
-            error={touchedFields.email ? errors.email?.message : undefined}
-            isValid={isValid("email")}
-          >
-            <input
-              id="email"
-              type="email"
-              autoComplete="email"
-              placeholder="you@example.com"
-              {...register("email")}
-              className={inputClass(
-                touchedFields.email && !!errors.email,
-                isValid("email")
-              )}
-            />
-          </FormField>
-
-          <FormField
-            label="Password"
-            id="password"
-            error={touchedFields.password ? errors.password?.message : undefined}
-            isValid={isValid("password")}
-          >
-            <input
-              id="password"
-              type="password"
-              autoComplete="current-password"
-              placeholder="••••••••"
-              {...register("password")}
-              className={inputClass(
-                touchedFields.password && !!errors.password,
-                isValid("password")
-              )}
-            />
-          </FormField>
-
-          <div className="flex justify-end">
-            <Link href="/auth/forgot-password" className="text-xs text-blue-600 hover:underline">
-              Forgot password?
-            </Link>
-          </div>
-
-          <button
-            type="submit"
-            disabled={isSubmitting}
-            className="flex items-center justify-center gap-2 rounded-lg bg-blue-600 py-2.5 text-sm font-medium text-white hover:bg-blue-700 disabled:opacity-60 transition-colors"
-          >
-            {isSubmitting && <Loader2 size={15} className="animate-spin" />}
-            Sign in
-          </button>
-        </form>
+        <LoginForm />
 
         <p className="mt-5 text-center text-sm text-gray-500">
           Don&apos;t have an account?{" "}
@@ -143,6 +39,17 @@ export default function LoginPage() {
         </p>
       </div>
     </div>
+  );
+}
+
+function GoogleIcon() {
+  return (
+    <svg width="18" height="18" viewBox="0 0 18 18" aria-hidden="true">
+      <path fill="#4285F4" d="M17.64 9.2c0-.637-.057-1.251-.164-1.84H9v3.481h4.844c-.209 1.125-.843 2.078-1.796 2.717v2.258h2.908c1.702-1.567 2.684-3.875 2.684-6.615z" />
+      <path fill="#34A853" d="M9 18c2.43 0 4.467-.806 5.956-2.184l-2.908-2.258c-.806.54-1.837.86-3.048.86-2.344 0-4.328-1.584-5.036-3.711H.957v2.332A8.997 8.997 0 0 0 9 18z" />
+      <path fill="#FBBC05" d="M3.964 10.707A5.41 5.41 0 0 1 3.682 9c0-.593.102-1.17.282-1.707V4.961H.957A8.996 8.996 0 0 0 0 9c0 1.452.348 2.827.957 4.039l3.007-2.332z" />
+      <path fill="#EA4335" d="M9 3.58c1.321 0 2.508.454 3.44 1.345l2.582-2.58C13.463.891 11.426 0 9 0A8.997 8.997 0 0 0 .957 4.961L3.964 7.293C4.672 5.163 6.656 3.58 9 3.58z" />
+    </svg>
   );
 }
